requirements.txt
streamlit==1.33.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0
scikit-learn==1.4.2
streamlit-lottie==0.0.5
langchain==0.1.14
langchain-community==0.0.33
langchain-core==0.1.35
langchain-groq==0.1.1
pdf2image==1.17.0
PyMuPDF==1.24.2
keybert==0.8.3
pyttsx3==2.90
pillow==10.3.0
faiss-cpu==1.8.0
